feat: create a library
